♻️ Migrate the domain model back to correct module.
