+++ gearmeshing_ai/mcp_client/base.py
@@ -15,8 +15,7 @@ import httpx
 
 from gearmeshing_ai.mcp_client.policy import PolicyMap, ToolPolicy
 from gearmeshing_ai.mcp_client.schemas.config import McpClientConfig
-from gearmeshing_ai.mcp_client.schemas.core import McpServerRef, McpTool, ToolCallResult
-from gearmeshing_ai.mcp_client.schemas.dto import ToolsPage
+from gearmeshing_ai.mcp_client.schemas.core import McpServerRef, McpTool, ToolCallResult, ToolsPage
 from gearmeshing_ai.mcp_client.strategy.base import is_mutating_tool_name
